[dev] moved localstorage check into utils
